# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Story Adventure Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /stories:
    get:
      operationId: listStories
      tags: [stories]
      summary: List story levels
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated story list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryListResponse"
    post:
      operationId: createStory
      tags: [stories]
      summary: Create a new story level (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryInput"
      responses:
        "201":
          description: Created story
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
  /stories/{id}:
    get:
      operationId: getStory
      tags: [stories]
      summary: Get a story level by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Story details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
    patch:
      operationId: updateStory
      tags: [stories]
      summary: Update a story level (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryUpdate"
      responses:
        "200":
          description: Updated story
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
    delete:
      operationId: deleteStory
      tags: [stories]
      summary: Delete a story level (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /stories/{id}/generate:
    post:
      operationId: generateStoryContent
      tags: [stories]
      summary: AI-generate story content for a level
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated story content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryContent"

  # ── GAMEPLAY ───────────────────────────────────────────────────────────────
...
  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get global leaderboard
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [alltime, weekly, monthly]
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
...
  /referrals/use:
    post:
      operationId: useReferralCode
      tags: [referrals]
      summary: Use a referral code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralCodeInput"
      responses:
        "200":
          description: Referral applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralResult"

  # ── PREMIUM ────────────────────────────────────────────────────────────────
...
  /withdrawals:
    get:
      operationId: listMyWithdrawals
      tags: [withdrawals]
      summary: List current user's withdrawal requests
      responses:
        "200":
          description: Withdrawals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Create a withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"

  # ── ADMIN ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ── USER ──────────────────────────────────────────────────────────────────
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        coins:
          type: integer
        gems:
          type: integer
        level:
          type: integer
        xp:
          type: integer
        role:
          type: string
          enum: [user, moderator, admin, superadmin]
        isPremium:
          type: boolean
        isBanned:
          type: boolean
        referralCode:
          type: string
        createdAt:
          type: string
      required: [id, clerkId, username, email, coins, gems, level, xp, role, isPremium, isBanned, referralCode, createdAt]
    UserProfileUpdate:
      type: object
      properties:
        username:
          type: string
        avatarUrl:
          type: string
...
    Story:
      type: object
      properties:
        id:
          type: integer
        levelNumber:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, legendary]
        coinsReward:
          type: integer
        gemsReward:
          type: integer
        xpReward:
          type: integer
        isPremium:
          type: boolean
        isPublished:
          type: boolean
        thumbnailUrl:
          type: ["string", "null"]
        estimatedMinutes:
          type: integer
        storyContent:
          type: ["string", "null"]
        choices:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/StoryChoice"
        createdAt:
          type: string
      required: [id, levelNumber, title, description, category, difficulty, coinsReward, gemsReward, xpReward, isPremium, isPublished, estimatedMinutes, createdAt]
    StoryChoice:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        consequence:
          type: string
      required: [id, text, consequence]
    StoryListResponse:
      type: object
      properties:
        stories:
          type: array
          items:
            $ref: "#/components/schemas/Story"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [stories, total, page, totalPages]
    StoryInput:
      type: object
      properties:
        levelNumber:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, legendary]
        coinsReward:
          type: integer
        gemsReward:
          type: integer
        xpReward:
          type: integer
        isPremium:
          type: boolean
        estimatedMinutes:
          type: integer
      required: [levelNumber, title, description, category, difficulty, coinsReward, gemsReward, xpReward, isPremium, estimatedMinutes]
    StoryUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, legendary]
        coinsReward:
          type: integer
        gemsReward:
          type: integer
        xpReward:
          type: integer
        isPremium:
          type: boolean
        isPublished:
          type: boolean
        estimatedMinutes:
          type: integer
    StoryContent:
      type: object
      properties:
        storyId:
          type: integer
        content:
          type: string
        choices:
          type: array
          items:
            $ref: "#/components/schemas/StoryChoice"
      required: [storyId, content, choices]

    # ── GAMEPLAY ──────────────────────────────────────────────────────────────
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: integer
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        score:
          type: integer
        level:
          type: integer
        isPremium:
          type: boolean
      required: [rank, userId, username, score, level, isPremium]
    LeaderboardResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
        type:
          type: string
        updatedAt:
          type: string
      required: [entries, type, updatedAt]
...
    ReferralCodeInput:
      type: object
      properties:
        code:
          type: string
      required: [code]
    ReferralResult:
      type: object
      properties:
        success:
          type: boolean
        coinsEarned:
          type: integer
        message:
          type: string
      required: [success, message]

    # ── PREMIUM ───────────────────────────────────────────────────────────────
...
    Withdrawal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        method:
          type: string
        accountDetails:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
        processedAt:
          type: ["string", "null"]
      required: [id, userId, amount, currency, method, accountDetails, status, createdAt]
    WithdrawalInput:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
        method:
          type: string
        accountDetails:
          type: string
      required: [amount, currency, method, accountDetails]

    # ── ADMIN ─────────────────────────────────────────────────────────────────
...